Split definition.go
